+++ packages/editor/src/index.ts
@@ -57,7 +57,6 @@ export type {
   EditableAPI,
   EditableAPIDeleteOptions,
   EditorChange,
-  EditorChanges,
   EditorSelection,
   EditorSelectionPoint,
   ErrorChange,
@@ -73,7 +72,6 @@ export type {
   OnPasteResultOrPromise,
   PasteData,
   PatchChange,
-  PatchObservable,
   RangeDecoration,
   RangeDecorationOnMovedDetails,
   ReadyChange,
