+++ packages/editor/tests/PortableTextEditor.test.tsx
@@ -3,7 +3,7 @@ import {createRef, type RefObject} from 'react'
 import {describe, expect, it, vi} from 'vitest'
 import type {EditorSelection} from '../src'
 import {PortableTextEditor} from '../src/editor/PortableTextEditor'
-import {InternalChange$Plugin} from '../src/plugins/plugin.internal.change-ref'
+import {InternalEditorChangePlugin} from '../src/plugins/plugin.internal.editor-change-ref'
 import {InternalPortableTextEditorRefPlugin} from '../src/plugins/plugin.internal.portable-text-editor-ref'
 import {createTestEditor} from '../src/test/vitest'
 
@@ -24,7 +24,7 @@ describe('initialization', () => {
     const {locator} = await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -46,7 +46,7 @@ describe('initialization', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -80,7 +80,7 @@ describe('initialization', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -121,7 +121,7 @@ describe('initialization', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -166,7 +166,7 @@ describe('initialization', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -211,7 +211,7 @@ describe('initialization', () => {
     const {editor} = await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
         </>
       ),
       initialValue: value,
@@ -289,7 +289,7 @@ describe('initialization', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
