+++ packages/editor/tests/valueNormalization.test.tsx
@@ -2,7 +2,7 @@ import {createRef, type RefObject} from 'react'
 import {describe, expect, it, vi} from 'vitest'
 import {defineSchema} from '../src'
 import {PortableTextEditor} from '../src/editor/PortableTextEditor'
-import {InternalChange$Plugin} from '../src/plugins/plugin.internal.change-ref'
+import {InternalEditorChangePlugin} from '../src/plugins/plugin.internal.editor-change-ref'
 import {InternalPortableTextEditorRefPlugin} from '../src/plugins/plugin.internal.portable-text-editor-ref'
 import {createTestEditor} from '../src/test/vitest'
 
@@ -29,7 +29,7 @@ describe('values: normalization', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <InternalEditorChangePlugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
