+++ packages/editor/tests/withPortableTextLists.test.tsx
@@ -1,7 +1,6 @@
 import {createRef, type RefObject} from 'react'
 import {describe, expect, it, vi} from 'vitest'
 import {PortableTextEditor} from '../src/editor/PortableTextEditor'
-import {InternalChange$Plugin} from '../src/plugins/plugin.internal.change-ref'
 import {InternalPortableTextEditorRefPlugin} from '../src/plugins/plugin.internal.portable-text-editor-ref'
 import {createTestEditor} from '../src/test/vitest'
 
@@ -40,12 +39,10 @@ describe('plugin:withPortableTextLists', () => {
         style: 'normal',
       },
     ]
-    const onChange = vi.fn()
 
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
