+++ packages/editor/tests/withPortableTextMarkModel.test.tsx
@@ -2,7 +2,7 @@ import {defineSchema} from '@portabletext/schema'
 import {createRef, type RefObject} from 'react'
 import {describe, expect, it, vi} from 'vitest'
 import {PortableTextEditor} from '../src/editor/PortableTextEditor'
-import {InternalChange$Plugin} from '../src/plugins/plugin.internal.change-ref'
+import {InternalEditorChangePlugin} from '../src/plugins/plugin.internal.editor-change-ref'
 import {InternalPortableTextEditorRefPlugin} from '../src/plugins/plugin.internal.portable-text-editor-ref'
 import {createTestEditor} from '../src/test/vitest'
 import type {EditorSelection} from '../src/types/editor'
@@ -44,7 +44,7 @@ describe('plugin:withPortableTextMarksModel', () => {
       await createTestEditor({
         children: (
           <>
-            <InternalChange$Plugin onChange={onChange} />
+            <InternalEditorChangePlugin onChange={onChange} />
             <InternalPortableTextEditorRefPlugin ref={editorRef} />
           </>
         ),
@@ -178,7 +178,7 @@ describe('plugin:withPortableTextMarksModel', () => {
       await createTestEditor({
         children: (
           <>
-            <InternalChange$Plugin onChange={onChange} />
+            <InternalEditorChangePlugin onChange={onChange} />
             <InternalPortableTextEditorRefPlugin ref={editorRef} />
           </>
         ),
@@ -285,7 +285,7 @@ describe('plugin:withPortableTextMarksModel', () => {
       await createTestEditor({
         children: (
           <>
-            <InternalChange$Plugin onChange={onChange} />
+            <InternalEditorChangePlugin onChange={onChange} />
             <InternalPortableTextEditorRefPlugin ref={editorRef} />
           </>
         ),
@@ -389,7 +389,7 @@ describe('plugin:withPortableTextMarksModel', () => {
       await createTestEditor({
         children: (
           <>
-            <InternalChange$Plugin onChange={onChange} />
+            <InternalEditorChangePlugin onChange={onChange} />
             <InternalPortableTextEditorRefPlugin ref={editorRef} />
           </>
         ),
@@ -460,7 +460,7 @@ describe('plugin:withPortableTextMarksModel', () => {
       await createTestEditor({
         children: (
           <>
-            <InternalChange$Plugin onChange={onChange} />
+            <InternalEditorChangePlugin onChange={onChange} />
             <InternalPortableTextEditorRefPlugin ref={editorRef} />
           </>
         ),
@@ -516,7 +516,7 @@ describe('plugin:withPortableTextMarksModel', () => {
       await createTestEditor({
         children: (
           <>
-            <InternalChange$Plugin onChange={onChange} />
+            <InternalEditorChangePlugin onChange={onChange} />
             <InternalPortableTextEditorRefPlugin ref={editorRef} />
           </>
         ),
@@ -566,7 +566,7 @@ describe('plugin:withPortableTextMarksModel', () => {
       await createTestEditor({
         children: (
           <>
-            <InternalChange$Plugin onChange={onChange} />
+            <InternalEditorChangePlugin onChange={onChange} />
             <InternalPortableTextEditorRefPlugin ref={editorRef} />
           </>
         ),
